# Office-management-railway-deployment/swagger.yaml
openapi: 3.0.0
info:
  title: Office Management API
  version: 1.0.0
  description: API documentation for the Office Management System Backend
servers:
  - url: http://localhost:3000
    description: Local server
  - url: https://office-management-sbuh.onrender.com
    description: Production server (Render)
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in a cookie
security:
  - cookieAuth: []
paths:
...
  /admin/emp/all:
    get:
      summary: Get all active employees
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of active employees
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
        "500":
          description: Server error
...
  /admin/leaves/all:
    get:
      summary: Get all leave requests
      tags:
        - Admin-Leaves
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of leave requests
        "500":
          description: Server error
...
  /employee/leaves/apply:
    post:
      summary: Apply for a new leave
      tags:
        - Employee-Leaves
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - start_date
                - end_date
                - reason
              properties:
                type:
                  type: string
                  enum:
                    - Sick
                    - Casual
                    - Privilege
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                reason:
                  type: string
      responses:
        "201":
          description: Leave application submitted successfully
        "400":
          description: Missing required fields or invalid leave type
        "409":
          description: Leave request already exists for these dates
        "500":
          description: Server error
  /employee/leaves/summary:
    get:
      summary: Get leave entitlement summary and balance
      tags:
        - Employee-Leaves
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of leave types with used and available balances
        "500":
          description: Server error
  /employee/leaves/all:
    get:
      summary: Get all leave applications for the current employee
      tags:
        - Employee-Leaves
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of all leaves
        "500":
          description: Server error
  /employee/notifications/all:
    get:
      summary: Get all notifications for the current employee
      tags:
        - Employee-Notifications
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of notifications
        "401":
          description: Unauthorized
        "500":
          description: Server error
  /employee/notifications/mark-read/all:
    patch:
      summary: Mark all notifications as read
      tags:
        - Employee-Notifications
      security:
        - cookieAuth: []
      responses:
        "200":
          description: All notifications marked as read
        "401":
          description: Unauthorized
        "500":
          description: Server error
  /employee/notifications/mark-read/{id}:
    patch:
      summary: Mark a specific notification as read
      tags:
        - Employee-Notifications
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
        "401":
          description: Unauthorized
        "404":
          description: Notification not found
        "500":
          description: Server error
  /employee/notifications/delete/{id}:
    delete:
      summary: Delete a specific notification
      tags:
        - Employee-Notifications
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification deleted successfully
        "401":
          description: Unauthorized
        "404":
          description: Notification not found
        "500":
          description: Server error
...
  /settings/change-password:
    post:
      summary: Change employee password
      tags:
        - Employee-Settings
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - oldPassword
                - newPassword
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        "200":
          description: Password updated successfully
        "401":
          description: Incorrect current password or unauthenticated
        "404":
          description: User not found
        "500":
          description: Server error
  /settings/updateProfile:
    put:
      summary: Update employee profile details
      tags:
        - Employee-Settings
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                location:
                  type: string
                bio:
                  type: string
                avatar:
                  type: string
                designation:
                  type: string
      responses:
        "200":
          description: Profile updated successfully
        "401":
          description: Not authenticated
        "404":
          description: User not found
        "500":
          description: Server error
...